updated the result component to include gautham changes
